upgrade to v2.24.0 (#582)
* Upgraded GeoServer from version 2.23.2 to 2.24.0 and reorder action to build image before testing

+++ scripts/start.sh
@@ -225,8 +225,10 @@ if [[ ${CLUSTERING} =~ [Tt][Rr][Uu][Ee] ]]; then
   fi
   # Download Clustering module, temporary fixes https://github.com/kartoza/docker-geoserver/issues/514
   ${request} https://download.jar-download.com/cache_jars/org.jdom/jdom2/2.0.6.1/jar_files.zip
-  unzip jar_files.zip -d  "${CATALINA_HOME}"/webapps/"${GEOSERVER_CONTEXT_ROOT}"/WEB-INF/lib/
-  rm jar_files.zip
+  if [[ -f jar_files.zip ]];then
+    unzip jar_files.zip -d  "${CATALINA_HOME}"/webapps/"${GEOSERVER_CONTEXT_ROOT}"/WEB-INF/lib/
+    rm jar_files.zip
+  fi
 fi
 
 export REQUEST_TIMEOUT PARALLEL_REQUEST GETMAP REQUEST_EXCEL SINGLE_USER GWC_REQUEST WPS_REQUEST
